Move schema specific code to schema visitor
