+++ packages/graphql-codegen-factories/src/schema/FactoriesSchemaVisitor.ts
@@ -6,9 +6,55 @@ import {
   InputObjectTypeDefinitionNode,
   InputValueDefinitionNode,
 } from "graphql";
-import { FactoriesBaseVisitor, TypeValue } from "../FactoriesBaseVisitor";
+import {
+  DeclarationBlock,
+  indent,
+} from "@graphql-codegen/visitor-plugin-common";
+import { FactoriesBaseVisitor } from "../FactoriesBaseVisitor";
+
+interface TypeValue {
+  defaultValue: string;
+  isNullable: boolean;
+}
 
 export class FactoriesSchemaVisitor extends FactoriesBaseVisitor {
+  protected convertObjectType(
+    node: ObjectTypeDefinitionNode | InputObjectTypeDefinitionNode
+  ): string {
+    return new DeclarationBlock(this._declarationBlockConfig)
+      .export()
+      .asKind("function")
+      .withName(
+        `${this.convertFactoryName(
+          node
+        )}(props: Partial<${this.convertNameWithNamespace(
+          node
+        )}>): ${this.convertNameWithNamespace(node)}`
+      )
+      .withBlock(
+        [
+          indent("return {"),
+          node.kind === "ObjectTypeDefinition"
+            ? indent(indent(`__typename: "${node.name.value}",`))
+            : null,
+          ...(node.fields ?? []),
+          indent(indent("...props,")),
+          indent("};"),
+        ]
+          .filter(Boolean)
+          .join("\n")
+      ).string;
+  }
+
+  protected convertField(
+    node: FieldDefinitionNode | InputValueDefinitionNode
+  ): string {
+    const { defaultValue, isNullable } = node.type as unknown as TypeValue;
+    return indent(
+      indent(`${node.name.value}: ${isNullable ? "null" : defaultValue},`)
+    );
+  }
+
   NamedType(node: NamedTypeNode): TypeValue {
     return {
       defaultValue: this.getDefaultValue(node.name.value),
